==> R/03_assign_taxa.R <==
#################################################
#
# Cross-taxa population monitoring bibliometrics
#
# 03_assign_taxa.R
#
# jan.perret@cefe.cnrs.fr
#################################################

#filter_taxa <- read.csv("./data/raw/mots_cles_taxons.csv", sep = ";", header = T)
#import du fichier csv contenant les mots clefs associes aux taxons
#attention pour l'instant il y a des majuscules dans le fichier

#creation des listes de taxons
# bird_list <- keywordsList(filter_taxa,"bird")
# mammal_list <- keywordsList(filter_taxa,"mammal")
# reptile_list <- keywordsList(filter_taxa,"reptile")
# amphibian_list <- keywordsList(filter_taxa,"amphibian")
# fish_list <- keywordsList(filter_taxa,"fish")
# invertebrate_list <- keywordsList(filter_taxa,"invertebrate")
# plant_list <- keywordsList(filter_taxa,"plant")
# fungi_list <- keywordsList(filter_taxa,"fungi")

# #assignation des oiseaux
# corpus <- taxaAssociation(corpus,column_names,bird_list,"bird")
# 
# #assignation des mammiferes
# corpus <- taxaAssociation(corpus,column_names,mammal_list,"mammal")
# 
# #assignation des reptiles
# corpus <- taxaAssociation(corpus,column_names,reptile_list,"reptile")
# 
# #assignation des amphibiens
# corpus <- taxaAssociation(corpus,column_names,amphibian_list,"amphibian")
# 
# #assignation des poissons
# corpus <- taxaAssociation(corpus,column_names,fish_list,"fish")
# 
# #assignation des invertebres
# corpus <- taxaAssociation(corpus,column_names,invertebrate_list,"invertebrate")
# 
# #assignation des plantes
# corpus <- taxaAssociation(corpus,column_names,plant_list,"plant")
# 
# #assignation des champignons
# corpus <- taxaAssociation(corpus,column_names,fungi_list,"fungi")
